seeds/chromadb/legislative-bill-seeding/requirements.txt
langchain
langchain-community
pypdf
sentence-transformers
langchain-huggingface
chromadb-client
PyYAML
langchain-chroma